% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup_admb.R
\name{cleanup_admb}
\alias{cleanup_admb}
\title{Cleans up environment after ADMB calls}
\usage{
cleanup_admb(folderName = NULL)
}
\arguments{
\item{folderName}{Character string: Name of the folder to be deleted (Default = NULL)}
}
\value{
Nothing
}
\description{
Removes left over files from a failed run and temporary files created by ADMB
}
\section{Files Removed}{


All files of type csv, txt, out, text are removed.
Temporary files "variance", "fmin.log", "eigv.rpt"
}

