% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_effort_folder_setup.R
\name{create_effort_folder_setup}
\alias{create_effort_folder_setup}
\title{create folders structure to hold model output.}
\usage{
create_effort_folder_setup(
  rootFolder,
  fleetNames,
  exploitationRates,
  create = T
)
}
\arguments{
\item{rootFolder}{Character string: name of folder to hold all output from a single set of parameters}

\item{exploitationRates}{Numeric matrix: Values for the exploitation rates for each fleet. Eg. Column 1 would be the values of exploitation rates from min to max for fllet 1, column 2 for fleet 2, etc.}

\item{create}{Boolean. To create folders on machine (TRUE), to retrun structure only (F)}

\item{fleetsNames}{Character vector: Names of the fleets in the model}
}
\value{
A Character matrix (3 columns):

\item{folderStructure}{Column 1 = hcrType, Column 2 = hcrLevels, column 3 = exploitationRate}
}
\description{
The folder structure is dependent on the number and type of scenarios selected in addition to the number of fleets. Every combination of fleet/exploitation is considered. If there are 3 fleets and 8 exploitation levels there will be 8x8x8 = 512 folders. Within each scenario are subfolder to store model output: dat and pin files and/or plots.
}
