% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_folder_setup.R
\name{create_folder_setup}
\alias{create_folder_setup}
\title{create folders structure to hold model output.}
\usage{
create_folder_setup(rootFolder, hcrTypes = c("Fixed", "Ramp"),
  hcrLevels = c("Complex", "HTSpecies", "LTSpecies"), exploitationRates)
}
\arguments{
\item{rootFolder}{Character string: name of folder to hold all output from a single set of parameters}

\item{hcrTypes}{Character vector: Names for the harvest control rule types (Default = c("Fixed, "Ramp"))}

\item{hcrLevels}{Character vector: Names for the levels that the harvest control rules are applied to. (Default = c("Complex","HTSpecies","LTSpecies"))}

\item{exploitationRates}{Numeric vector: Values for the maximum exploitation rate in any scenario.}
}
\value{
Nothing is returned
}
\description{
The folder structure is dependent on the number and type of scenarios selected. Within each scenario are subfolder to store model output
dat and pin files and/or plots. Additional folders are also created ofr summary statistics and plots.
}
