% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darwin.R
\name{darwin}
\alias{darwin}
\title{Performs Darwinian process}
\usage{
darwin()
}
\description{
Simulates and selects parameter values that result in model runs resembling historical patterns.
The parameters to be simulated can be selected by the user
}
