% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_to_Rlist.R
\name{file_to_Rlist}
\alias{file_to_Rlist}
\title{Reads output from a single Hydra into an R list}
\usage{
file_to_Rlist(fn)
}
\arguments{
\item{fn}{character. Filename of output to read}
}
\value{
List of variable length
\item{indices}{Each list item is an index from the model run where the name of the list item is the name of the index as specified in the outputfile}
}
\description{
Reads in a model run output file of variable length and formats it into a List where each list item is the name of
an index calculated in the model. THIS NEED TO BE CODED IN C++ TO SPEED THINGS UP
}
