% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydramse.r
\docType{package}
\name{hydramse}
\alias{hydramse}
\alias{hydramse-package}
\title{hydramse: Implementation of MSE process}
\description{
Core files used to set up and execute an MSE process
}
\details{
\itemize{
\item Create folder structure for model runs
}

To learn more about using \code{hydramse}, start with the vignette: \code{browseVignettes(package="hydramse")} or click \code{index} link below
}
