% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lengthen_hydra_data.R
\name{lengthen_hydra_data}
\alias{lengthen_hydra_data}
\title{lengthen hydra data for additional years}
\usage{
lengthen_hydra_data(data, nYrs)
}
\arguments{
\item{data}{A datafile resembling \code{hydraDataList}}

\item{nYrs}{The total number of years Hydra should be extended to}
}
\value{
List
\item{Nyrs}{Scalar. The Input value of nYrs}
\item{recruitmentCov}{Vector. Lengthened vector for recruitment covariate. Uses mean of current values}
\item{maturityCov}{Vector. Lengthened vector for maturity covariate. Uses mean of current values}
\item{growthCov}{Vector. Lengthened vector for growth covariate. Uses mean of current values}
\item{observedTemperature}{Matrix. Lengthened vector for observed temperature. Uses mean of current values}
\item{observedBiomass}{Matrix. Lengthened vector for observed Biomass. Uses mean of current values}
\item{observedCatch}{Matrix. Lengthened vector for observed Catch. Uses mean of current values}
\item{observedEffort}{Matrix. Lengthened vector for observed Effort. Uses 0 for fist part of time series}
\item{redundantRecDevs}{Matrix. Lengthened vector for observed Effort. Uses 0 for fist part of time series}
}
\description{
The Darwinian study utilizes Hydra but requires a longer period of time to run
Here we extend all o fthe variables passed to Hydra to make up the additional years
The first nYrs of the model run is assumed to be a period without fishing pressure
followed by a period (of length \code{hydraDataList$Nyrs}) of fishing.
}
\section{Notes}{

The 8 variables that are time dependent in the model are listed above.
We need to provide dummy data for the first x years during the zero fishing period
where we assume the model will reach some kind of equilibrium
}

