% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_box_whiskers.R
\name{plot_box_whiskers}
\alias{plot_box_whiskers}
\title{creates box whisker plots}
\usage{
plot_box_whiskers(
  filePath,
  rootFolder,
  inputFile,
  dataType,
  scaling = T,
  outFilename = NULL
)
}
\arguments{
\item{filePath}{Character string. Path to rootfolder}

\item{rootFolder}{Character string. Folder containing all runs output}

\item{inputFile}{Character string. RDS file to process.}

\item{dataType}{Character string. Used for yaxis label}

\item{scaling}{binary scalar. Scale to kilotons (T) or not (F)}

\item{outFilename}{Character string. Name of file to be saved. Default = inputFile.png}
}
\value{
Plots are written
}
\description{
Multipanel plots representing species/guild biomass, catch, etc over differnt scenario types
}
