% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_darwinRules.R
\name{plot_darwinRules}
\alias{plot_darwinRules}
\title{Plots Stock recruitment model}
\usage{
plot_darwinRules(data = darwinRules)
}
\description{
Uses parameter values found in lazy loaded data \code{darwinRules}. These models are considered starting points for simulation study
}
