% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_production_function_guild.R
\name{plot_production_function_guild}
\alias{plot_production_function_guild}
\title{plots production functions}
\usage{
plot_production_function_guild(filePath, rootFolder, dataType,
  scaling = T, nLastYrs = 20)
}
\arguments{
\item{filePath}{Character string. Path to rootfolder}

\item{rootFolder}{Character string. Folder containing all runs output}

\item{dataType}{Character string. Used for yaxis label}

\item{scaling}{binary scalar. Scale to kilotons (T) or not (F)}

\item{nYrs}{numeric scalar. number of trailing years to use for calculations. Default = 20 (Last 20 yrs of sim)}
}
\value{
Figures are written to rootFolder (1 figure per Guild)
}
\description{
multipanel plots the production function for each guild /scenario combination
}
