% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_radar_scenarios.R
\name{plot_radar_scenarios}
\alias{plot_radar_scenarios}
\title{creates Radar plots for scenario comparison}
\usage{
plot_radar_scenarios(filePath, rootFolder, inputFile, timePeriod,
  indicesNames)
}
\arguments{
\item{filePath}{Character string. Path to rootfolder}

\item{rootFolder}{Character string. Folder containing all runs output}

\item{inputFile}{Character string. RDS file to process.}

\item{timePeriod}{List. Time frame to average over. "early" and "late" fields eg (early = 21:30)}

\item{indices}{Character vector. The names of the indices of interest from the model run.}
}
\value{
Plots are written
}
\description{
Multipanel plots representing scenario/Exploitation rate across several
variables used to measure ecosystem health
}
