% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_darwinData.R
\name{plot_stock_recruit}
\alias{plot_stock_recruit}
\title{Plots Stock recruitment model}
\usage{
plot_stock_recruit(data = darwinData, stockRecruitData)
}
\arguments{
\item{data}{List. lazy data \code{darwinData} is default}
}
\description{
Uses parameter values found in lazy loaded data \code{darwinRules}. These models are considered starting points for simulation study
FINISH LATER
}
