% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_time_series.r
\name{plot_time_series}
\alias{plot_time_series}
\title{Plots times eries data}
\usage{
plot_time_series(dataset)
}
\arguments{
\item{dataset}{character. Either "discards", "landings", or "survey"}
}
\description{
Plots landings, discards, survey data (lazy loaded data \code{timeSeries}).
}
