% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_single_scenario.R
\name{process_single_scenario}
\alias{process_single_scenario}
\title{Process all runs for a given scenario}
\usage{
process_single_scenario(modelOutput, indices, revenueData,
  convertBiomassTolbs = 2204.62)
}
\arguments{
\item{modelOutput}{List (length = nSims) content of files}

\item{indices}{Character vector. Names of the indices to process}

\item{revenueData}{Data frame. External revenue data}

\item{convertBiomassTolbs}{Numeric scalar. 1 metric tonne = 2204.62 lbs (default)}
}
\value{
A list containing the following items:
\item{species_bio}{biomass by species}
\item{species_catch}{catch by species}
}
\description{
A scenario is defined a a harvest control rule with a maximum exploitation rate.
Each scenario is run a number of times. All of the runs for a scenario are processed for each index
For biomass and catch indices at the species and the guild level the median is taken over all runs at each time point
Fo othe indices, median at each time point is calculated.
}
