% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules_set.R
\name{rule1_biomass}
\alias{rule1_biomass}
\title{Assess if model passes 1st biomass rule}
\usage{
rule1_biomass(biomass, nYrsFishing, historicBounds, simulationRules,
  speciesList)
}
\arguments{
\item{biomass}{matrix. (nSpecies x nYrs) average biomass (avByr variable)}

\item{nYrsFishing}{scalar. Number of fishing years simulated}

\item{historicBounds}{matrix. (nSpecies x ??) from lazy data \code{darwinData}}

\item{simulationRules}{list. adapted from lazy data  \code{darwinRules}}

\item{speciesList}{Character vector. Names of species as defined in lazily loaded data set}
}
\value{
List
\item{lowerBoolean}{boolean vector (nSpecies). Denotes which species populations were above historic lower bounds.}
\item{pass}{boolean value. Indicates whether all species passed the test}
}
\description{
Check to ensure all species population are above a historic minimum
This minimum is the minimum of survey data and multispecies model output
}
