% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules_set.R
\name{rule2_biomass}
\alias{rule2_biomass}
\title{Assess if model passes second biomass rule.}
\usage{
rule2_biomass(biomass, nYrs, historicBounds, simulationRules)
}
\arguments{
\item{biomass}{matrix. (nSpecies x nYrs) average biomass (avByr variable)}

\item{nYrs}{scalar. Total number of years in simulation}

\item{historicBounds}{matrix. (nSpecies x ??) from lazy data \code{darwinData}}

\item{simulationRules}{list. adapted from lazy data  \code{darwinRules}}
}
\value{
List
\item{lowerBoolean}{boolean vector (nSpecies). Denotes which species populations were above historic lower bounds.}
\item{upperBoolean}{boolean vector (nSpecies). Denotes which species populations were below historic upper bounds.}
\item{pass}{boolean value. Indicates whether all species passed the test}
}
\description{
After a period of fishing pressure the species population are checked to ensure they fall withing historic bounds
}
