% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules_set.R
\name{rule3_landings}
\alias{rule3_landings}
\title{Assess if model passes catch rule.}
\usage{
rule3_landings(catch, historicBounds, simulationRules, speciesList)
}
\arguments{
\item{historicBounds}{matrix. (nSpecies x ??) from lazy data \code{darwinData}}

\item{simulationRules}{list. adapted from lazy data  \code{darwinRules}}

\item{speciesList}{Character vector. Names of species as defined in lazily loaded data set}

\item{biomass}{matrix. (nSpecies x nYrs) average biomass (avByr variable)}
}
\value{
List
\item{lowerBoolean}{boolean vector (nSpecies). Denotes which species populations were above historic lower bounds.}
\item{upperBoolean}{boolean vector (nSpecies). Denotes which species populations were below historic upper bounds.}
\item{pass}{boolean value. Indicates whether all species passed the test}
}
\description{
After a period of fishing pressure the catch of each species are checked to ensure they fall withing historic bounds
}
