% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_single_hydra.R
\name{run_single_hydra}
\alias{run_single_hydra}
\title{Executes Hydra ADMB code}
\usage{
run_single_hydra(n = 1, exePath, datPath, pinPath)
}
\arguments{
\item{n}{Numeric scalar. The seed used for the model run (Default: n = 1)}

\item{exePath}{Character string. The path to the executable}

\item{datPath}{Character string. The path to the dat file}

\item{pinPath}{Character string. The path to the pin file}
}
\value{
Nothing is returned. Output is written to a file for processing
}
\description{
Runs the Hydra code
}
