% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_parameters.R
\name{simulate_parameters}
\alias{simulate_parameters}
\title{simulate all unknowns in the MSE}
\usage{
simulate_parameters(stockRecruitParams, simulationRules)
}
\arguments{
\item{simulationRules}{These have been lazily loaded as darwinRules}
}
\value{
List
\item{otherFood}{simulated value of otherFood variable}
\item{alphaShepherd}{simulated value of alpha for the shepherd model  }
\item{betaShepherd}{simulated value of beta for the shepherd model}
\item{shapeShepherd}{value of the shape parameter. Predefined by user in darwinRules$SRType}
\item{alphaHockey}{}
\item{betaHockey}{}
\item{shapeHockey}{}
\item{alphaSegmented}{}
\item{alphaSegmented}{}
\item{alphaSegmented}{}
}
\description{
Currently only simulate stock recruitment parameters and other food. See Method section below for more details
}
\section{Method}{

Methods are described in the vignette \code{simulation_documentation}
}

