% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_hydra_data.R
\name{update_hydra_data}
\alias{update_hydra_data}
\title{update base Hydra data}
\usage{
update_hydra_data(hydraD, updates)
}
\arguments{
\item{hydraD}{List of current base data (as in \code{hydradata::hydradataList})}

\item{updates}{List containing data with fieldnames identical to those in \code{hydradata::hydradataList} used to update the base data}
}
\value{
List
\item{updatedData}{List of \code{hydraD} with updated values}
}
\description{
update Hydra data with values stored in updates
}
