% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_box_whiskers.R
\name{plot_box_whiskers}
\alias{plot_box_whiskers}
\title{creates box whisker plots}
\usage{
plot_box_whiskers(filePath, rootFolder, inputFile)
}
\arguments{
\item{filePath}{Character string. Path to rootfolder}

\item{rootFolder}{Character string. Folder containing all runs output}

\item{inputFile}{Character string. RDS file to process.}
}
\value{
Plots are written
}
\description{
Multipanel plots representing species/guild biomass, catch, etc over differnt scenario types
}
