% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stock_recruit.R
\name{plot_stock_recruit}
\alias{plot_stock_recruit}
\title{Plots Stock recruitment model}
\usage{
plot_stock_recruit(data = darwinData, stockRecruitData)
}
\arguments{
\item{data}{List. lazy data \code{darwinData} is default}

\item{stockRecruitData}{NOT FINISHED}
}
\description{
Plots SR models found in lazy loaded data \code{darwinRules}.
Overlays with simulated stock recruitment models
}
