% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_darwin_output.R
\name{process_darwin_output}
\alias{process_darwin_output}
\title{processes output from Darwin run}
\usage{
process_darwin_output(filePath)
}
\arguments{
\item{filePath}{Character string. Path to where output files exist}
}
\value{
A List:
\item{biomass}{matrix (nSpecies x nYrs). Average biomass (avByr) for each species in each year}
\item{catch}{matrix (nSpecies x nYrs). Average catch (est_catch_biomass) for each species in each year}
}
\description{
Reads in hydra output from the specified darwin run
}
