% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_model_runs.r
\name{process_model_runs}
\alias{process_model_runs}
\title{Process model runs}
\usage{
process_model_runs(data, indices, scenarios, rootFolder,
  outPutScenarioDirs, revenueData, outputType = "indices", nLastYrs)
}
\arguments{
\item{data}{List. The hydra dataList \code{hydradata::hydraDataList}}

\item{indices}{Character vector. The names of the indices of interest from the model run.}

\item{scenarios}{Character vector. The names of the scenario. Defined as a harvest control rul with a max exploitation rate}

\item{rootFolder}{root folder holding all run info}

\item{outPutScenarioDirs}{list of scenario directories under the rootFolder}

\item{revenueData}{Data frame (see \code{\link{process_single_scenario}})}

\item{outputType}{Character string. Name of folder where Hydra output files are stored ("indices","diagnostics")}

\item{nLastYrs}{Numeric scalar. Number of trailing years in each run to take mean over. Default = 20 (Last 20 yrs of sim)}
}
\value{
Nothing. RDS files are produced
}
\description{
Reads in all of the model run data, processes it, and writes to an .rds file. Details of the processing can be found in section below.
Contents of rds file are tidy and used in plotting routines
}
\section{Processing details}{


A model run (scenario) is considered to be n simulations of the Hydra model. Stochasticity differentiates each simulation.

Each scenario/ model run is processed as in \code{\link{process_single_scenario}}. All of the processed scenarios and the relevant indices are then exported
as an RDS file
}

