% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_model_runs.r
\name{process_model_runs}
\alias{process_model_runs}
\title{Process model runs}
\usage{
process_model_runs(data, indices, scenarios, exRateVec, rootFolder,
  outPutScenarioDirs, revenueData, outputType = "indices")
}
\arguments{
\item{data}{List. The hydra dataList \code{hydradata::hydraDataList}}

\item{indices}{Character vector. The names of the indices of interest from the model run.}

\item{scenarios}{Character vector. The names of the scenario. Defined as a harvest control rul with a max exploitation rate}

\item{exRateVec}{Numeric vector. Exploitation rates used in the MSE}

\item{rootFolder}{root folder holding all run info}

\item{outPutScenarioDirs}{list of scenario directories under the rootFolder}

\item{revenueData}{Data frame (see \code{\link{process_single_scenario}})}
}
\value{
Nothing
}
\description{
Reads in all of the model run data, processes it, and writes to an .rdata file. Details of the processing can be found in section below
}
\section{Processing details}{


Each scenario run ... more coming
}

